move providers to seperate application
